+++ gems/aws-sdk-s3/lib/aws-sdk-s3/encryptionV2/kms_cipher_provider.rb
@@ -50,7 +50,7 @@ module Aws
         # @return [Cipher] Given an encryption envelope, returns a
         #   decryption cipher.
         def decryption_cipher(envelope, options = {})
-          encryption_context = Json.load(envelope['x-amz-matdesc'])
+          encryption_context = extract_encryption_context(envelope['x-amz-matdesc'])
           cek_alg = envelope['x-amz-cek-alg']
 
           case envelope['x-amz-wrap-alg']
@@ -114,6 +114,13 @@ module Aws
         end
 
         private
+
+        # Raise a decryption error for a malformed material description.
+        def extract_encryption_context(matdesc)
+          Json.load(matdesc)
+        rescue Aws::Json::ParseError, EncodingError
+          raise Errors::DecryptionError, 'Malformed material description'
+        end
 
         def validate_key_wrap(key_wrap_schema)
           case key_wrap_schema